# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for live stream configuration, ingest/publish lifecycle, DVR recordings, and event hooks.",
        "title": "Open Streamer API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/hooks": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hooks"
                ],
                "summary": "List hooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.HookList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hooks"
                ],
                "summary": "Create hook",
                "parameters": [
                    {
                        "description": "Hook configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Hook"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/apidocs.HookData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
        "/hooks/{hid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hooks"
                ],
                "summary": "Get hook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hook ID",
                        "name": "hid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.HookData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hooks"
                ],
                "summary": "Update hook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hook ID",
                        "name": "hid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Hook configuration",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Hook"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.HookData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "hooks"
                ],
                "summary": "Delete hook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hook ID",
                        "name": "hid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
        "/hooks/{hid}/test": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hooks"
                ],
                "summary": "Test hook delivery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hook ID",
                        "name": "hid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidocs.HookTestData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/apidocs.ErrorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apidocs.ErrorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "apidocs.HookData": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.Hook"
                }
            }
        },
        "apidocs.HookList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Hook"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "apidocs.HookTestData": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/apidocs.HookTestInner"
                }
            }
        },
        "apidocs.HookTestInner": {
            "type": "object",
            "properties": {
                "at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "domain.EventType": {
            "type": "string",
            "enum": [
                "session.opened",
                "session.closed",
                "stream.created",
                "stream.updated",
                "stream.started",
                "stream.stopped",
                "stream.deleted",
                "input.connected",
                "input.reconnecting",
                "input.degraded",
                "input.failed",
                "input.failover",
                "input.recovered",
                "recording.started",
                "recording.stopped",
                "recording.failed",
                "segment.written",
                "dvr.segment_pruned",
                "transcoder.started",
                "transcoder.stopped",
                "transcoder.error",
                "push.started",
                "push.active",
                "push.reconnecting",
                "push.failed",
                "config.changed",
                "watermark.asset_created",
                "watermark.asset_deleted",
                "hook.created",
                "hook.updated",
                "hook.deleted"
            ],
            "x-enum-comments": {
                "EventDVRSegmentPruned": "retention loop deleted an aged-out segment",
                "EventInputConnected": "source connected successfully",
                "EventInputDegraded": "error detected by manager",
                "EventInputFailed": "worker exited / non-retriable",
                "EventInputFailover": "switched to a different input",
                "EventInputReconnecting": "transient error, retrying",
                "EventInputRecovered": "previously degraded input is healthy again (auto or via failback probe)",
                "EventPushActive": "handshake ok; media flowing",
                "EventPushFailed": "dest.Limit exhausted or terminal error",
                "EventPushReconnecting": "write fail or input discontinuity, retrying",
                "EventPushStarted": "dialing target, pre-handshake",
                "EventStreamUpdated": "PUT /streams/{code} on existing record"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "PUT /streams/{code} on existing record",
                "",
                "",
                "",
                "source connected successfully",
                "transient error, retrying",
                "error detected by manager",
                "worker exited / non-retriable",
                "switched to a different input",
                "previously degraded input is healthy again (auto or via failback probe)",
                "",
                "",
                "",
                "",
                "retention loop deleted an aged-out segment",
                "",
                "",
                "",
                "dialing target, pre-handshake",
                "handshake ok; media flowing",
                "write fail or input discontinuity, retrying",
                "dest.Limit exhausted or terminal error",
                "",
                "",
                "",
                "",
                "",
                ""
            ],
            "x-enum-varnames": [
                "EventSessionOpened",
                "EventSessionClosed",
                "EventStreamCreated",
                "EventStreamUpdated",
                "EventStreamStarted",
                "EventStreamStopped",
                "EventStreamDeleted",
                "EventInputConnected",
                "EventInputReconnecting",
                "EventInputDegraded",
                "EventInputFailed",
                "EventInputFailover",
                "EventInputRecovered",
                "EventRecordingStarted",
                "EventRecordingStopped",
                "EventRecordingFailed",
                "EventSegmentWritten",
                "EventDVRSegmentPruned",
                "EventTranscoderStarted",
                "EventTranscoderStopped",
                "EventTranscoderError",
                "EventPushStarted",
                "EventPushActive",
                "EventPushReconnecting",
                "EventPushFailed",
                "EventConfigChanged",
                "EventWatermarkAssetCreated",
                "EventWatermarkAssetDeleted",
                "EventHookCreated",
                "EventHookUpdated",
                "EventHookDeleted"
            ]
        },
...
        "domain.Hook": {
            "type": "object",
            "properties": {
                "batch_flush_interval_sec": {
                    "description": "BatchFlushIntervalSec is the maximum time a batch may sit before being\nflushed even when below BatchMaxItems. 0 = use HooksConfig default,\nthen DefaultHookBatchFlushIntervalSec.",
                    "type": "integer"
                },
                "batch_max_items": {
                    "description": "BatchMaxItems caps the number of events bundled into one HTTP POST\nbody. 0 = use HooksConfig.BatchMaxItems, then DefaultHookBatchMaxItems.\nIgnored for File hooks (they always write one event per line).",
                    "type": "integer"
                },
                "batch_max_queue_items": {
                    "description": "BatchMaxQueueItems caps the per-hook in-memory queue (pending +\nre-queued failures). When exceeded, the OLDEST events are dropped\nwith a warning log so the queue never grows unbounded against an\nunreachable target. 0 = use HooksConfig default, then\nDefaultHookBatchMaxQueueItems.",
                    "type": "integer"
                },
                "enabled": {
                    "type": "boolean"
                },
                "event_types": {
                    "description": "EventTypes filters which events trigger delivery. Empty = all events.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.EventType"
                    }
                },
                "id": {
                    "type": "string"
                },
                "max_retries": {
                    "description": "MaxRetries is the number of delivery attempts before giving up.\n0 means use the server default (3). For HTTP hooks this caps retries\ninside a single batch flush; events that still fail are re-queued\nfor the next flush regardless of MaxRetries.",
                    "type": "integer"
                },
                "metadata": {
                    "description": "Metadata holds user-defined key-value pairs merged into every event payload\ndelivered by this hook. Useful for tagging events with custom context\n(e.g. environment, tenant ID, region) without modifying the server config.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "secret": {
                    "description": "HMAC-SHA256 signing secret (HTTP only)",
                    "type": "string"
                },
                "stream_codes": {
                    "description": "StreamCodes filters delivery by stream code.\nOnly and Except are mutually exclusive; Only takes precedence when both are set.\nOmitting the field (nil) means all streams are included.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.StreamCodeFilter"
                        }
                    ]
                },
                "target": {
                    "description": "HTTP(S) URL or absolute file path",
                    "type": "string"
                },
                "timeout_sec": {
                    "description": "TimeoutSec is the per-attempt delivery timeout in seconds.\n0 means use the server default (10s).",
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/domain.HookType"
                }
            }
        },
        "domain.HookType": {
            "type": "string",
            "enum": [
                "http",
                "file"
            ],
            "x-enum-varnames": [
                "HookTypeHTTP",
                "HookTypeFile"
            ]
        },
...
        "domain.StreamCodeFilter": {
            "type": "object",
            "properties": {
                "except": {
                    "description": "Except delivers events for all streams except those in this list.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "only": {
                    "description": "Only delivers events only for streams in this list.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}